# feat: improve performance when detecting country codes