--- a/lib/phonelib/core.rb
+++ b/lib/phonelib/core.rb
@@ -10,6 +10,11 @@ module Phonelib
       @@phone_data ||= load_data.freeze
     end
 
+    # @private getter for phone data indexed by country code (internal use only)
+    def data_by_country_codes
+      @@data_by_country_codes ||= phone_data.each_value.group_by { |d| d[COUNTRY_CODE] }
+    end
+
     # @private used to cache frequently-used regular expressions
     @@phone_regexp_cache = {}
 
@@ -179,7 +184,7 @@ module Phonelib
     def add_additional_regex(country, type, national_regex)
       return unless Phonelib::Core::TYPES_DESC.keys.include?(type.to_sym)
       return unless national_regex.is_a?(String)
-      @@phone_data = nil
+      @@phone_data = @@data_by_country_codes = nil
       @@additional_regexes[country.to_s.upcase] ||= {}
       @@additional_regexes[country.to_s.upcase][type] ||= []
       @@additional_regexes[country.to_s.upcase][type] << national_regex