--- a/lib/phonelib/phone_analyzer.rb
+++ b/lib/phonelib/phone_analyzer.rb
@@ -124,15 +124,22 @@ module Phonelib
     #
     # * +phone+ - phone number for parsing
     def detect_and_parse(phone, country)
-      result = {}
-      Phonelib.phone_data.each do |key, data|
+      phone = phone.gsub(/^00/, '')
+      countries_data = country_code_candidates_for(phone).flat_map { |code|
+        Phonelib.data_by_country_codes[code] || []
+      }
+      countries_data.each_with_object({}) { |data, result|
+        key = data[:id]
         parsed = parse_single_country(phone, data)
         if (!Phonelib.strict_double_prefix_check || key == country) && double_prefix_allowed?(data, phone, parsed && parsed[key])
           parsed = parse_single_country(changed_dp_phone(key, phone), data)
         end
         result.merge!(parsed) unless parsed.nil?
-      end
-      result
+      }.compact
+    end
+
+    def country_code_candidates_for(phone)
+      (1..3).map { |length| phone[0, length] }
     end
 
     # Create phone representation in e164 format