--- a/phonelib.gemspec
+++ b/phonelib.gemspec
@@ -30,6 +30,7 @@ Gem::Specification.new do |s|
   s.add_development_dependency 'codeclimate-test-reporter', '~> 1.0.0'
   s.add_development_dependency 'simplecov'
   s.add_development_dependency 'benchmark-ips'
+  s.add_development_dependency 'benchmark-memory'
   # fixing CI tests
   s.add_development_dependency 'rack-cache', '= 1.2'
   s.add_development_dependency 'json', '= 2.3.1'