--- a/spec/benchmark_helper.rb
+++ b/spec/benchmark_helper.rb
@@ -3,6 +3,7 @@
 ENV["RAILS_ENV"] ||= 'test'
 
 require 'benchmark/ips'
+require 'benchmark/memory'
 require 'benchmark'
 require 'phonelib'
 require 'rspec'