… Gemfile.lock
… bin/rspec